+++ packages/cds-web-overlays/src/navigation/__stories__/NavigationBar.stories.tsx
@@ -13,11 +13,11 @@ export default {
 
 const a11ySkipConfig = {
   config: {
-    /** The NavigationBar docs explain the proper way to setup the tabpanel */
-    rules: [
-      { id: 'aria-valid-attr-value', enabled: false },
-      { id: 'aria-required-children', enabled: false },
-    ],
+    /**
+     * The TabNavigation docs explain the proper way to setup the tabpanel
+     * @link https://cds.cbhq.net/components/tab-navigation#accessibility
+     */
+    rules: [{ id: 'aria-valid-attr-value', enabled: false }],
   },
 };