+++ packages/cds-web-overlays/src/navigation/__stories__/NavigationStorySetup.tsx
@@ -81,9 +81,7 @@ type NavigationBarFullExampleProps = {
   onTabChange?: (id: string) => void;
   onBackPress?: () => void;
 };
-/**
- * @deprecated this component will be removed from cds-web Q22023. It has been moved to cds-web-overlays.
- */
+
 export function NavigationBarFullExample({
   pageTitle,
   onTabChange,
@@ -140,9 +138,6 @@ export function NavigationBarFullExample({
   );
 }
 
-/**
- * @deprecated this component will be removed from cds-web Q22023. It has been moved to cds-web-overlays.
- */
 export const NavigationBarTitle = () => {
   return (
     <NavigationBar
@@ -162,9 +157,6 @@ export const NavigationBarTitle = () => {
   );
 };
 
-/**
- * @deprecated this component will be removed from cds-web Q22023. It has been moved to cds-web-overlays.
- */
 export const ComposedSystem = () => {
   const [activeSidebarIndex, setActiveSidebarIndex] = useState(0);
   const [activeTabId, setActiveTabId] = useState('all');
@@ -227,9 +219,6 @@ const testOverrideClass = css`
   }
 `;
 
-/**
- * @deprecated this component will be removed from cds-web Q22023. It has been moved to cds-web-overlays.
- */
 export const NavLinkExample = () => {
   const [activeItem, setActiveItem] = useState('assets');
   const getProps = (name: string) => ({
@@ -280,9 +269,6 @@ type SidebarExampleProps = {
   children?: React.ReactNode;
 } & Omit<SidebarMoreMenuProps, 'children'>;
 
-/**
- * @deprecated this component will be removed from cds-web Q22023. It has been moved to cds-web-overlays.
- */
 export const SidebarExample = ({ children, ...props }: SidebarExampleProps) => {
   const [activeIndex, setActiveIndex] = useState(0);
   const [moreMenuValue, setMoreMenuValue] = useState<string | undefined>(undefined);