+++ packages/web/dropdown/__stories__/DropdownInteractive.stories.tsx
@@ -56,14 +56,6 @@ export const SimpleDropdown: ComponentStoryObj<typeof Default> = {
   },
 };
 
-SimpleDropdown.parameters = {
-  a11y: {
-    config: {
-      rules: [{ id: 'aria-required-children', enabled: false }],
-    },
-  },
-};
-
 // Make sure opening dropdown doesn't scroll the page. This will be captured in percy.
 export const ScrollContainer: ComponentStoryObj<typeof Default> = {
   args: {
@@ -73,14 +65,5 @@ export const ScrollContainer: ComponentStoryObj<typeof Default> = {
     const canvas = within(canvasElement);
     // open the menu
     await userEvent.click(canvas.getByText('Open Menu'));
-  },
-};
-
-// Copying this approach from SimpleDropdown
-ScrollContainer.parameters = {
-  a11y: {
-    config: {
-      rules: [{ id: 'aria-required-children', enabled: false }],
-    },
   },
 };